refactor(checker): extract spread-overload-selection helpers to sibling module (2000-LOC cap)
Pure relocation to clear the checker 2000-LOC architecture cap that the
TS2556 fix pushed two files over (overload_resolution.rs 2021,
candidate_collection.rs 2128).

- New sibling `spread_overload_selection.rs`: the open-ended-spread
  classification (`call_has_open_ended_array_spread_argument`), the
  spread positional-index helper, the overload non-rest-skip decision
  (`skip_non_rest_overloads_for_open_ended_spread`, replacing the inline
  block at the call site), and the cohesive `find_prior_non_tuple_spread_
  for_mismatch` spread helper — all moved verbatim.
- `recheck_overload_args_after_mismatch_without_context` and
  `arg_source_span` moved into the existing `diagnostics.rs` sibling
  (speculative-call diagnostic recovery).
- `type_param_variadic_tuple_spread` widened to `pub(super)` for reuse.

No behavior change: byte-identical helper bodies, same call sites.
overload_resolution.rs 1971, candidate_collection.rs 1925 (both < 2000).
Build + clippy --all-targets clean; arch_guard.py passes; repro stays
clean; the overload/spread/rest/2556 nextest families pass with the same
two pre-existing env-only failures as HEAD (net-new 0).
